refactored structure a bit
