+++ components/common/svg/DarkThemeIcon.tsx
@@ -1,7 +1,7 @@
 import { Text } from 'react-native';
 
-import { colors } from '../../config';
-import { Theme } from '../../redux';
+import { colors } from '../../../config';
+import { Theme } from '../../../redux';
 
 const DarkThemeIcon = () => (
   <Text style={{ color: colors[Theme.LIGHT].color }}>🌘</Text>
