+++ components/common/svg/LightThemeIcon.tsx
@@ -1,7 +1,7 @@
 import { Text } from 'react-native';
 
-import { colors } from '../../config';
-import { Theme } from '../../redux';
+import { colors } from '../../../config';
+import { Theme } from '../../../redux';
 
 const LightThemeIcon = () => (
   <Text style={{ color: colors[Theme.DARK].color }}>🌞</Text>
